Print Summary Stats for a previous BCF run This function is built using the coda package and meant to mimick output from Stan. It includes:

* summary statistics for MCMC * effective sample sizes for each parameter * Gelman and Rubin's convergence diagnostics for each parameter

summarise_bcf(bcf_out)

Arguments

bcf_out
output from a BCF predict run

Examples

# \donttest{

# data generating process
p = 3 #two control variables and one moderator
n = 250

set.seed(1)

x = matrix(rnorm(n*p), nrow=n)

# create targeted selection
q = -1*(x[,1]>(x[,2])) + 1*(x[,1]<(x[,2]))

# generate treatment variable
pi = pnorm(q)
z = rbinom(n,1,pi)

# tau is the true (homogeneous) treatment effect
tau = (0.5*(x[,3] > -3/4) + 0.25*(x[,3] > 0) + 0.25*(x[,3]>3/4))

# generate the response using q, tau and z
mu = (q + tau*z)

# set the noise level relative to the expected mean function of Y
sigma = diff(range(q + tau*pi))/8

# draw the response variable with additive error
y = mu + sigma*rnorm(n)

# If you didn't know pi, you would estimate it here
pihat = pnorm(q)

bcf_fit = bcf(y, z, x, x, pihat, nburn=2000, nsim=2000)

# Get model fit diagnostics
summarise_bcf(bcf_fit)

#> Summary statistics for each Markov Chain Monte Carlo run

#> 
#> Iterations = 1:2000
#> Thinning interval = 1 
#> Number of chains = 4 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>              Mean      SD  Naive SE Time-series SE
#> sigma     0.34645 0.01795 0.0002007      0.0003487
#> tau_bar   0.47802 0.05860 0.0006551      0.0014099
#> mu_bar    0.07486 0.03890 0.0004349      0.0009617
#> yhat_bar  0.32879 0.02190 0.0002448      0.0002448
#> mu_scale  0.74997 0.17638 0.0019719      0.0290651
#> b0       -0.05806 0.07659 0.0008563      0.0054298
#> b1        0.32507 0.09796 0.0010952      0.0077435
#> 
#> 2. Quantiles for each variable:
#> 
#>               2.5%     25%      50%       75%   97.5%
#> sigma     0.313215  0.3340  0.34599  0.358117 0.38300
#> tau_bar   0.362250  0.4379  0.47839  0.517414 0.59213
#> mu_bar   -0.002933  0.0491  0.07501  0.101182 0.15104
#> yhat_bar  0.285457  0.3140  0.32890  0.343472 0.37201
#> mu_scale  0.479852  0.6106  0.72668  0.864408 1.15249
#> b0       -0.233721 -0.1011 -0.05118 -0.005843 0.07177
#> b1        0.157754  0.2565  0.31724  0.384626 0.54238
#> 
#> 
#> ----
#> 

#> Effective sample size for each parameter

#>      sigma    tau_bar     mu_bar   yhat_bar   mu_scale         b0         b1 
#> 2803.81525 1780.54993 1781.73824 8000.00000   42.51471  195.26432  188.67934 
#> 
#> ----
#> 

#> Gelman and Rubin's convergence diagnostic for each parameter

#> Potential scale reduction factors:
#> 
#>          Point est. Upper C.I.
#> sigma          1.01       1.02
#> tau_bar        1.00       1.00
#> mu_bar         1.00       1.00
#> yhat_bar       1.00       1.00
#> mu_scale       1.18       1.48
#> b0             1.05       1.12
#> b1             1.09       1.25
#> 
#> Multivariate psrf
#> 
#> 1.14

# }