Print posterior summary stats and MCMC diagnostics for a fitted BCF object. This function is built using the coda package and meant to mimic output from rstan::print.stanfit(). It includes, for key parameters, posterior summary stats, effective sample sizes, and Gelman and Rubin's convergence diagnostics.

Print posterior summary stats and MCMC diagnostics for a fitted BCF object. This function is built using the coda package and meant to mimic output from rstan::print.stanfit(). It includes, for key parameters, posterior summary stats, effective sample sizes, and Gelman and Rubin's convergence diagnostics.

summarise_bcf(bcf_out, params_2_summarise = c("sigma", "tau_bar",
  "mu_bar", "yhat_bar"))

Arguments

bcf_out
output from a BCF predict run

Examples

# \donttest{

# data generating process
p = 3 #two control variables and one moderator
n = 250

set.seed(1)

x = matrix(rnorm(n*p), nrow=n)

# create targeted selection
q = -1*(x[,1]>(x[,2])) + 1*(x[,1]<(x[,2]))

# generate treatment variable
pi = pnorm(q)
z = rbinom(n,1,pi)

# tau is the true (homogeneous) treatment effect
tau = (0.5*(x[,3] > -3/4) + 0.25*(x[,3] > 0) + 0.25*(x[,3]>3/4))

# generate the response using q, tau and z
mu = (q + tau*z)

# set the noise level relative to the expected mean function of Y
sigma = diff(range(q + tau*pi))/8

# draw the response variable with additive error
y = mu + sigma*rnorm(n)

# If you didn't know pi, you would estimate it here
pihat = pnorm(q)

bcf_fit = bcf(y, z, x, x, pihat, nburn=2000, nsim=2000)

# Get model fit diagnostics
summarise_bcf(bcf_fit)

#> Summary statistics for each Markov Chain Monte Carlo run

#> 
#> Iterations = 1:2000
#> Thinning interval = 1 
#> Number of chains = 4 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>             Mean      SD  Naive SE Time-series SE
#> sigma    0.34624 0.01801 0.0002013      0.0004095
#> tau_bar  0.47980 0.05851 0.0006541      0.0012103
#> mu_bar   0.07359 0.03840 0.0004294      0.0008018
#> yhat_bar 0.32871 0.02200 0.0002460      0.0002405
#> 
#> 2. Quantiles for each variable:
#> 
#>               2.5%     25%     50%     75%  97.5%
#> sigma    0.3129667 0.33383 0.34562 0.35821 0.3832
#> tau_bar  0.3648250 0.44082 0.47964 0.51910 0.5941
#> mu_bar   0.0002023 0.04776 0.07335 0.09917 0.1502
#> yhat_bar 0.2855042 0.31388 0.32879 0.34345 0.3723
#> 
#> 
#> ----
#> 

#> Effective sample size for each parameter

#>    sigma  tau_bar   mu_bar yhat_bar 
#> 2240.084 2421.469 2308.988 8379.601 
#> 
#> ----
#> 

#> Gelman and Rubin's convergence diagnostic for each parameter

#> Potential scale reduction factors:
#> 
#>          Point est. Upper C.I.
#> sigma          1.01       1.03
#> tau_bar        1.00       1.00
#> mu_bar         1.00       1.00
#> yhat_bar       1.00       1.00
#> 
#> Multivariate psrf
#> 
#> 1.01
#> 
#> ----
#> 

# }